+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Added
 - documents conversion
 - ignoring self signed certificate 
+- Added connection to a demo document server
 - Document Editing Service address is now splitted in two settings: inner address (address that confluence will use to access service) and public address (address that user will use to access editors)
 - validation of server settings on the settings page
 - ignoring self signed certificate